Separated ConsoleReader from ConstoleController
